requirements.dev.txt
black
django-tailwind[reload]

# Data science experiments
numpy
matplotlib
umap-learn
tqdm
